refactor: invitation-management.tsx をコンポーネント単位で分割、同伴者名のアコーディオン表示
